Remove crypto-browserify polyfill from host
The host build aliased bare `crypto` to crypto-browserify solely to satisfy
one static `import { createHmac } from 'crypto'` in
runtime-common/url-signature.ts — a Node-only signing path the host never
executes (host uses the WebCrypto `createURLSignature` path).

Split url-signature.ts so the browser path stays import-free of crypto, and
move the synchronous Node path (`createURLSignatureSync`/`verifyURLSignature`)
into url-signature-node.ts using `node:crypto`. Realm-server's two consumers
import from the new module. Drop the Vite `crypto` alias, the host dependency,
and the catalog entry — removing crypto-browserify and 34 transitive packages.
