+++ packages/host/package.json
@@ -102,7 +102,6 @@
     "camelcase": "catalog:",
     "concurrently": "catalog:",
     "content-tag": "catalog:",
-    "crypto-browserify": "catalog:",
     "date-fns": "catalog:",
     "decorator-transforms": "catalog:",
     "ember-animated": "catalog:",
