+++ packages/runtime-common/url-signature.ts
@@ -5,12 +5,11 @@
  * The signature is HMAC-SHA256(token, urlPath) where:
  * - token is used as the HMAC key
  * - urlPath is the pathname + search params (without the sig param)
+ *
+ * This module holds only the Web Crypto (browser) signing path, so it pulls in
+ * no Node builtins and is safe to bundle for the browser. The synchronous
+ * Node-only signing/verification path lives in `./url-signature-node`.
  */
-
-// HMAC for the Node-only signing path below. Bare `crypto` resolves to the
-// Node builtin under native Node, and to crypto-browserify in the host build
-// via its Vite alias (host browsers only reach the Web Crypto path above).
-import { createHmac } from 'crypto';
 
 // Browser implementation using Web Crypto API
 export async function createURLSignature(
@@ -40,26 +39,4 @@ export async function createURLSignature(
     .replace(/\+/g, '-')
     .replace(/\//g, '_')
     .replace(/=+$/, ''); // URL-safe base64
-}
-
-// Node.js implementation
-export function createURLSignatureSync(token: string, url: URL): string {
-  let urlForSigning = new URL(url.href);
-  urlForSigning.searchParams.delete('sig');
-
-  let message = urlForSigning.pathname + urlForSigning.search;
-  let signature = createHmac('sha256', token)
-    .update(message)
-    .digest('base64url');
-
-  return signature;
-}
-
-export function verifyURLSignature(
-  token: string,
-  url: URL,
-  providedSignature: string,
-): boolean {
-  let expectedSignature = createURLSignatureSync(token, url);
-  return expectedSignature === providedSignature;
 }
