+++ pnpm-workspace.yaml
@@ -126,7 +126,6 @@ catalog:
   content-tag: ^4.0.0
   countries-list: ^3.1.1
   cron: ^3.1.7
-  crypto-browserify: ^3.12.0
   date-fns: ^2.28.0
   dayjs: ^1.11.7
   decorator-transforms: ^2.3.0
